fix: Revert non-impl files to named exports and rename Webiny folder to ProjectAws
