+++ extensions/_wip/idp/MyIdpExtension.tsx
@@ -1,12 +1,10 @@
 import React from "react";
 import { Api } from "webiny/extensions";
 
-const MyIdpExtension = () => {
+export const MyIdpExtension = () => {
     return (
         <>
             <Api.Extension src={"/extensions/_wip/idp/MyIdp.ts"} />
         </>
     );
 };
-
-export default MyIdpExtension;
