+++ extensions/graphql/MySchemaExtension.tsx
@@ -1,12 +1,10 @@
 import React from "react";
 import { Api } from "webiny/extensions";
 
-const MySchemaExtension = () => {
+export const MySchemaExtension = () => {
     return (
         <>
             <Api.Extension src={"/extensions/graphql/MyGraphQLSchema.ts"} />
         </>
     );
 };
-
-export default MySchemaExtension;
