+++ packages/project-aws/src/extensions/ProjectAws/AutoInstall.ts
