+++ packages/project-aws/src/extensions/ProjectAws/AutoInstall/AutoInstallAfterApiDeploy.ts
