+++ packages/project-aws/src/extensions/ProjectAws/SetAdminEnvVars/SetAdminEnvVarsBeforeWatch.ts
