+++ packages/project-aws/src/extensions/ProjectAws/SetDatabaseSetupOutput.ts
