+++ packages/project-aws/src/extensions/index.ts
@@ -1,4 +1,4 @@
 export * from "./OpenSearch.js";
 export * from "./AwsDefaultRegion.js";
-export * from "./Webiny/AutoInstall.js";
+export * from "./ProjectAws/AutoInstall.js";
 export * from "./ProjectAws.js";
